Move some files around and continue POC on MainScreen to prove out performance improvements
